requirements.txt
Flask==3.0.0
requests==2.31.0
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
matplotlib==3.8.2  # Optional, for future visualizations
python-dotenv==1.0.0  # For environment variables